+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - [#694](https://github.com/estruyf/vscode-front-matter/issues/694): Start terminal session from the folder where the `frontmatter.json` file is located
 - [#696](https://github.com/estruyf/vscode-front-matter/issues/696): Close the local server terminal on restart
 - [#699](https://github.com/estruyf/vscode-front-matter/issues/699): Changing border theme variable for the dashboard header
+- [#703](https://github.com/estruyf/vscode-front-matter/issues/703): Fix retrieval of Astro Collections for `pnpm` projects
 
 ## [9.3.1] - 2023-10-27